Refactor tests to use PostgresSQL and setup shared test configuration
Tests have been refactored to use PostgresSQL instead of MySQL, a 'TestConfiguration' file is also added to centralize and manage the configurations shared across all tests. This change is to ensure that tests are more consistent and maintainable.
Some unused test files and methods were also removed to improve the code cleanliness, and Regex in protocols files were modified to match the correct parameters.
Furthermore, a Postgres restore script is added to assist in managing the database state for tests.